+++ extension/reader/boot.js
@@ -1,0 +1,1 @@
+window.__SPEEDREAD_EXTENSION__ = true;